mutare clase in fisiere separate
