+++ Image.java
@@ -1,0 +1,12 @@
+class Image implements Element {
+    private String url;
+
+    public Image(String url) {
+        this.url = url;
+    }
+
+    @Override
+    public void print() {
+        System.out.println("Image: " + url);
+    }
+}
