+++ Paragraph.java
@@ -1,0 +1,12 @@
+class Paragraph implements Element {
+    private String text;
+
+    public Paragraph(String text) {
+        this.text = text;
+    }
+
+    @Override
+    public void print() {
+        System.out.println("Paragraph: " + text);
+    }
+}
