+++ Section.java
@@ -1,0 +1,33 @@
+import java.util.ArrayList;
+import java.util.List;
+class Section implements Element {
+    private String title;
+    private List<Element> children = new ArrayList<>();
+
+    public Section(String title) {
+        this.title = title;
+    }
+
+    @Override
+    public void add(Element e) {
+        children.add(e);
+    }
+
+    @Override
+    public void remove(Element e) {
+        children.remove(e);
+    }
+
+    @Override
+    public Element get(int index) {
+        return children.get(index);
+    }
+
+    @Override
+    public void print() {
+        System.out.println("Section: " + title);
+        for (Element child : children) {
+            child.print();
+        }
+    }
+}
