+++ Table.java
@@ -1,0 +1,12 @@
+class Table implements Element {
+    private String something;
+
+    public Table(String something) {
+        this.something = something;
+    }
+
+    @Override
+    public void print() {
+        System.out.println("Table: " + something);
+    }
+}
