+++ lab3.java
@@ -1,132 +1,3 @@
-import java.util.ArrayList;
-import java.util.List;
-
-interface Element {
-    void print();
-    default void add(Element e) {}
-    default void remove(Element e) {}
-    default Element get(int index) { return null; }
-}
-
-class Paragraph implements Element {
-    private String text;
-
-    public Paragraph(String text) {
-        this.text = text;
-    }
-
-    @Override
-    public void print() {
-        System.out.println("Paragraph: " + text);
-    }
-}
-
-class Image implements Element {
-    private String url;
-
-    public Image(String url) {
-        this.url = url;
-    }
-
-    @Override
-    public void print() {
-        System.out.println("Image: " + url);
-    }
-}
-
-class Table implements Element {
-    private String something;
-
-    public Table(String something) {
-        this.something = something;
-    }
-
-    @Override
-    public void print() {
-        System.out.println("Table: " + something);
-    }
-}
-
-class Section implements Element {
-    private String title;
-    private List<Element> children = new ArrayList<>();
-
-    public Section(String title) {
-        this.title = title;
-    }
-
-    @Override
-    public void add(Element e) {
-        children.add(e);
-    }
-
-    @Override
-    public void remove(Element e) {
-        children.remove(e);
-    }
-
-    @Override
-    public Element get(int index) {
-        return children.get(index);
-    }
-
-    @Override
-    public void print() {
-        System.out.println("Section: " + title);
-        for (Element child : children) {
-            child.print();
-        }
-    }
-}
-
-class Book implements Element {
-    private String title;
-    private ArrayList<Author> authors = new ArrayList<>();
-    private List<Element> contents = new ArrayList<>();
-
-    public Book(String title) {
-        this.title = title;
-    }
-
-    public void addAuthor(Author author) {
-        this.authors.add(author);
-    }
-
-    @Override
-    public void add(Element e) {
-        contents.add(e);
-    }
-
-    @Override
-    public void print() {
-        System.out.println("Book: " + title);
-        if (authors != null) {
-            for (Author author : authors) {
-                author.print();
-            }
-        }
-        for (Element content : contents) {
-            content.print();
-        }
-    }
-}
-
-class Author {
-    private String name;
-    private String surname;
-
-    public Author(String fullName) {
-        String[] names = fullName.split(" ");
-        this.name = names[0];
-        this.surname = names[1];
-    }
-
-    public void print() {
-        System.out.println("Author: " + name + " " + surname);
-    }
-}
-
-
 public class lab3 {
     public static void main(String[] args) {
         Book noapteBuna = new Book("Noapte buna, copii!");
